pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-heatmap"
version = "0.1.0"
description = "Analyze how AI coding agents navigate and understand codebases"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click",
    "gitpython",
    "rich",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pre-commit",
    "pytest",
    "ruff",
    "types-setuptools",
]

[project.scripts]
agent-heatmap = "agent_heatmap.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
exclude = [
    "build",
    "dist",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["agent_heatmap"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
